Ajout calcul de stats

+++ src/main/java/fournisseur/behaviors/CreationCatalogueBehavior.java
@@ -101,8 +101,8 @@ public class CreationCatalogueBehavior extends OneShotBehaviour {
             for (Integer numProduit : listProduitFournisseur) {
                 getDataStore().put(listProduit.get(numProduit), 0);
             }
-
-            getDataStore().put("Tresorerie", new Double(10000));// set du montant initial de pesos
+            
+            getDataStore().put("Tresorerie", new Double(100000));// set du montant initial de pesos
         } catch (ParseException ex) {
             Logger.getLogger(CreationCatalogueBehavior.class.getName()).log(Level.SEVERE, "Format de message BDD incorrect");
         }
